making new surface code backwards compatible with previous examples
Issue was that previous examples would not run with new surface code without going back and changing them to specify zero surface fluxes. Now changed so that iceberg fluxes, plume fluxes and surface fluxes will all be assumed to be 0 if nothing is specified about each of them. That is, when running fjordrpm, you only need to specify the active processes, and the inactive processes will be assumed to be inactive and will not crash the code.