fixing section that adjusts layer thicknesses to coincide with sill depth
Noticed a feature whereby the layer thickness were being adjusted to coincide with sill depth in such a way that you got uniform resolution above and below the sill. This is unhelpful if you are trying to run with layer thicknesses that are refined in some region (e.g., near the surface). Adjusted so that the layers are squeezed/stretched by a constant factor above and below the sill so that the non-uniformity in input layer thicknesses is preserved. Also required a few extra checks on the inputs.