input profile interpolation fix
Changed the interpolation of the shelf T/S profiles. Previously this used pchip for both interpolation and extrapolation, but this could create silly results (e.g., if CTD forcing profile was only to 400 m, then the model did an extrapolation to 800 m, you could get silly values). Now the interpolation is pchip, but the extrapolation is done using nearest, so if the model needs values deeper than the CTD it will just fill the bottom value constantly down, and similarly if the CTD profile does not go quite to the surface, the model will fill the shallowest CTD value to the surface.